+++ tests/more-templates.test.js
@@ -6,7 +6,7 @@ test('imdb', t => {
   let str = `{{IMDb title | 0426883 | Alpha Dog }}`;
   var doc = wtf(str);
   let obj = doc.templates(0);
-  t.equal(obj.template, 'imdb', 'imdb');
+  t.equal(obj.template, 'imdb title', 'imdb');
   t.equal(obj.id, '0426883', 'id1');
   t.equal(obj.title, 'Alpha Dog', 'title');
